refactor: extract agent-task config/policy types into leaf module
The agent_task_schedule file mixed ~16 plain config/policy data types (schedule
options, retry/rotation policy, resolved provider policy, output bindings,
execution budgets, adoption records) inside a private `config` submodule with the
scheduler runtime that depends on core::plan and core::agent_task. Those config
types are self-contained (no scheduler/plan deps) and are the shared contract
surface consumed by the scheduler, dispatch service, and the lab-contract type
layer alike.

Move the config submodule into its own leaf core::agent_task_config module
(pulling the four config-only serde-default helpers with it) and re-export it
from agent_task_schedule (`pub use agent_task_config::*`) so all existing
agent_task_schedule::* call sites stay stable.

This isolates the policy types the lab-contract crate needs from the heavy
plan -> gate -> proof scheduler chain. Cut 4a toward the homeboy-lab-contract
foundation crate.
